+++ src/model/asset.rs
@@ -1,0 +1,5 @@
+#[derive(Clone)]
+pub struct Asset {
+    /// identifier of asset, formatted as asset_name#domain_id
+    pub id: String,
+}
