+++ src/model/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod accounts;
 pub mod assets;
 pub mod domains;
+pub mod oob;
 pub mod peers;
