+++ tests/assets.rs
@@ -1,24 +1,34 @@
 #[cfg(test)]
 mod tests {
-
-    use async_std::task;
-    use iroha::{model, storage::kura};
+    use iroha::{
+        model::{
+            block::Block,
+            commands::{
+                accounts::{CreateAccount, CreateRole},
+                assets::{CreateAsset, TransferAsset},
+                domains::CreateDomain,
+                oob::Command,
+            },
+            tx::Transaction,
+        },
+        storage::kura,
+    };
 
     #[async_std::test]
     async fn transfer_asset_from_account1_to_account2() {
-        let create_role_command = model::Command {
+        let create_role_command = Command {
             version: 1,
             command_type: 0,
-            payload: model::commands::accounts::CreateRole {
+            payload: CreateRole {
                 role_name: "user".to_string(),
                 permissions: Vec::new(),
             }
             .into(),
         };
-        let create_domain_command = model::Command {
+        let create_domain_command = Command {
             version: 1,
             command_type: 0,
-            payload: model::commands::domains::CreateDomain {
+            payload: CreateDomain {
                 domain_id: "domain".to_string(),
                 default_role: "user".to_string(),
             }
@@ -26,37 +36,37 @@ mod tests {
         };
         let account1_id = "account1_name@domain".to_string();
         let account2_id = "account2_name@domain".to_string();
-        let create_account1_command = model::Command {
+        let create_account1_command = Command {
             version: 1,
             command_type: 0,
-            payload: model::commands::accounts::CreateAccount {
+            payload: CreateAccount {
                 account_name: "account1_name".to_string(),
                 domain_id: "domain".to_string(),
                 public_key: [63; 32],
             }
             .into(),
         };
-        let create_account2_command = model::Command {
+        let create_account2_command = Command {
             version: 1,
             command_type: 0,
-            payload: model::commands::accounts::CreateAccount {
+            payload: CreateAccount {
                 account_name: "account2_name".to_string(),
                 domain_id: "domain".to_string(),
                 public_key: [63; 32],
             }
             .into(),
         };
-        let create_asset_command = model::Command {
+        let create_asset_command = Command {
             version: 1,
             command_type: 0,
-            payload: model::commands::assets::CreateAsset {
+            payload: CreateAsset {
                 asset_name: "xor".to_string(),
                 domain_id: "domain".to_string(),
                 precision: 0,
             }
             .into(),
         };
-        let transfer_asset_command: model::Command = model::commands::assets::TransferAsset {
+        let transfer_asset_command: Command = TransferAsset {
             source_account_id: account1_id.clone(),
             destination_account_id: account2_id.clone(),
             asset_id: "xor".to_string(),
@@ -64,11 +74,11 @@ mod tests {
             amount: 200.2,
         }
         .into();
-        let block = model::Block {
+        let block = Block {
             height: 0,
             timestamp: 0,
             transactions: vec![
-                model::Transaction {
+                Transaction {
                     commands: vec![
                         create_role_command,
                         create_domain_command,
@@ -81,7 +91,7 @@ mod tests {
                     quorum: 1,
                     signatures: vec![],
                 },
-                model::Transaction {
+                Transaction {
                     commands: vec![transfer_asset_command],
                     creation_time: 1,
                     account_id: "source@domain".to_string(),
